Fix possible ValueError in ObjectSerializer.php
ObjectSerializer.php behaviour is based on type of the content received. However, as of now the content types are handled incorrectly. 

Line 336 contains types "object", "mixed" and "\DateTime" that are handled before, and thus it's unnecessary to handle them here.

The same line 336 contains a "\SplFileObject" that is handled later, and it requires special handling, it must not be passed to settype(). If "\SplFileObject" is passed to settype() (line 337), a ValueError will be thrown. 

Therefore this is not just a supeficial change.